improvement: svm dtos in their own module
